Diploma Supplement Bug Fix

+++ src/main/java/org/fenixedu/academic/domain/serviceRequests/IDiplomaSupplementRequest.java
@@ -36,7 +36,7 @@ public interface IDiplomaSupplementRequest extends IDocumentRequest {
 
     public String getPrevailingScientificArea(final Locale locale);
 
-    public long getEctsCredits();
+    public double getEctsCredits();
 
     public DegreeOfficialPublication getDegreeOfficialPublication();
